The Phonon of Search
This patch adds support for two additional storage and query backends.
One is already functional, the other is a placeholder at the moment.
Functional the HyperEstaier backend. This is very fast and low-level,
but remarably easy to add. The query language is different from what
we're currently using and a query translater should be added later.

The second backend is Xapian. It's present in the build system, it
compiles, but that's it for the moment.



svn path=/branches/work/kde4/playground/libs/archivereader/; revision=547331