changed string to float for weight. These typeURIs should be respected by indexers soon.
Added and optional flag to control indexing process of identifiers chemistry.inchi and chemistry.molecular_formula.
Tokenized=false  should be respected by indexwriters and processed by the search engine backends in the way that it will not be split into tokens (by default  Tokenized=true). This is especially important for identifiers, which contain special characters and/or space delimiters.

svn path=/trunk/kdesupport/strigi/; revision=685348