Cygwin fixes by Ben van Klinken
For fun i tried compiling strigi with cygwin. I got the core things
working (deepfind, etc), here's a small patch of the things i had to
change to
make it work (against the current svn).
Mostly compile stuff (assumptions are made that if the compiler is not
msvc then it is mingw). Also there are some ambiguous casts for
addValue. And
a small typo in the strigiconfig.h.cmake file (which wasn't actually
necessary, but is probably incorrect).


svn path=/trunk/kdesupport/strigi/; revision=704666