Added Xesam User Language parser. Now it will be possible to handle Xesam UserLanguage queries (http://wiki.freedesktop.org/wiki/XesamUserSearchLanguage).
This is the first working version of the parser. It's quite functional but there's still some work do to.
The parser relies on bison. If it isn't available on the system Strigi will use default generated files committed to svn.

With both Xesam userlanguage and querylanguage parsers done, Xesam support is full.

svn path=/trunk/kdesupport/strigi/; revision=707885