The function 'getFiles(int depth) in Strigi scales very badly. It will return all files at a certain depth. For large indices this can be millions of files. To fix this situation, we replaced getFiles(int depth) with getChildren(const string& parenturi). In this way, the list of files can be retrieved in well defined bits. An additional advantage is that the files are retrieved by directory of archive file. This will make the indexing algorithm traverse the files in a more logical and piecemeal manner.
We needed to add another function to AnalysisResult, namely parentPath(), which retrieves the path of the parent of a particular file. We chose the name 'parent' instead of 'directory' because a parent can also be e.g. a zip file.

The changes required for this are quite a few and at the time of this patch not all changes have been finalized. This means indexing and search may stop workig for you for a few days until the biggest problems are ironed out. In addition, you should remove your index in ~/.strigi. The index as generated by the last svn version did not contain all information that is essential for Strigi to operate. This error was introduced by switching to a new metadata format.


svn path=/trunk/kdesupport/strigi/; revision=716530