Fix a bug in subinputstream.
Under certain circumstances the function read() of the internal stream could be 
called with max < min. read() specifies that in such cases, there is no limit on the number of bytes that may be read. This would cause SubInputStream to malfunction because it would allow too much of the internal stream to be read.


svn path=/trunk/kdesupport/strigi/; revision=814834