Move Strigi::DirLister to ArchiveReader::DirLister
The class Strigi::DirLister currently occurrs twice in Strigi. This
accidental circumstance is undesirable and hence the less important
class is moved. Strigi::Dirlister in archivereader.h is not used outside
of Strigi currently. Therefore, while this change is not binary
compatible, it poses not problems with current binaries using
libstreams. Strigi::DirLister is moved to
Strigi::ArchiveReader::DirLister. This makes sense since DirLister is a
simple helper class for iterating through entries in a directory that
ArchiveReader is reading.

A new Strigi release will follow in a couple of days unless problems are spotted
with this change.

svn path=/trunk/kdesupport/strigi/; revision=908973