Rewrite the implementation of ArchiveReader
The new implementation is more effient in listing contents of
directories. Now single directory entries can be returned without the
need for reading the entire archive of which the directory is a part.
This is very useful for GUI applications which want to interactively
list the contents of archives.

The implementation is rather tricky and has spent quite a long time in a
local git repository. It is now good enough for addition to the Strigi
trunk.

svn path=/trunk/kdesupport/strigi/; revision=928707