Fix race issue that limits performance.
FileLister::nextDir could return -1 even though the listing was not finished yet, because one thread had popped a dir off the stack and was analyzing it. The analysis would yield another dir, but in the meantime FileLister::nextDir returns -1. This is now fixed by keeping the mutex locked until the list with directories has become non-empty or the analysis has really finished.

svn path=/trunk/kdesupport/strigi/; revision=1066944