Another go at solving the Snow Leopard compilation issue.
It seems that on Snow Leopard (x86_64) uint_64_t is unsigned long long, while
dbus_uint64_t is unsigned long. This leads to a missing method overload, as
described here, amongst other places: http://trac.macports.org/ticket/21138.
After some discussion with Jos, it seems best to re-define uint64_t to match
the dbus type, locally, and on OSX only, since it would be much harder to
change the code generator to use dbus types, and the header can't be fixed
to have an additional overload without platform ifdefs or static_assert
hackery to figure out if the types are the same. Not pretty, but such are
things here in the cross platform real world ;).

svn path=/trunk/kdesupport/strigi/; revision=1068179