Improve cpp analyzer.
This is a pretty poor analyzer but less so now.
A bug is fixes where it would read every text file completely hoping for an '#include' statement anywhere. Now it checks if the mimetype is correct (this is the bad part, but better than before) and requires every '#include' to be preceded with only whitespace.

svn path=/trunk/kdesupport/strigi/; revision=1075837